refactor: split template.ts into schema-types, utils, and render modules
The 450-line template.ts mixed type definitions, string utilities, and
the TypeRenderer class. Split into three focused modules:
- schema-types.ts — HerokuSchema, SchemaNode, and related interfaces
- utils.ts — pure string utilities (toPascalCase, renderJSDoc, etc.)
- render.ts — TypeRenderer class and its private helpers

template.ts is now a barrel re-export so all existing imports continue
to work unchanged.
